[arcane,utils] Déplace le test de 'Event' dans un test unitaire utilisant 'gtest'.
