+++ arcane/src/arcane/tests/UtilsUnitTest.cc
@@ -1,11 +1,11 @@
 ﻿// -*- tab-width: 2; indent-tabs-mode: nil; coding: utf-8-with-signature -*-
 //-----------------------------------------------------------------------------
-// Copyright 2000-2024 CEA (www.cea.fr) IFPEN (www.ifpenergiesnouvelles.com)
+// Copyright 2000-2025 CEA (www.cea.fr) IFPEN (www.ifpenergiesnouvelles.com)
 // See the top-level COPYRIGHT file for details.
 // SPDX-License-Identifier: Apache-2.0
 //-----------------------------------------------------------------------------
 /*---------------------------------------------------------------------------*/
-/* UtilsUnitTest.cc                                            (C) 2000-2024 */
+/* UtilsUnitTest.cc                                            (C) 2000-2025 */
 /*                                                                           */
 /* Test des fonctions utilitaires de Arcane.                                 */
 /*---------------------------------------------------------------------------*/
@@ -16,7 +16,6 @@
 #include "arcane/utils/ArgumentException.h"
 #include "arcane/utils/AutoDestroyUserData.h"
 #include "arcane/utils/HPReal.h"
-#include "arcane/utils/IFunctorWithArgument.h"
 #include "arcane/utils/IMemoryInfo.h"
 #include "arcane/utils/ITraceMng.h"
 #include "arcane/utils/PlatformUtils.h"
@@ -33,18 +32,10 @@
 #include "arcane/core/FactoryService.h"
 #include "arcane/core/ServiceBuilder.h"
 
-#include "arcane/tests/ArcaneTestGlobal.h"
-
 #include <fenv.h>
 
 /*---------------------------------------------------------------------------*/
 /*---------------------------------------------------------------------------*/
-
-namespace Arcane
-{
-extern "C++" ARCANE_UTILS_EXPORT void
-_internalTestEvent();
-}
 
 namespace ArcaneTest
 {
@@ -119,7 +110,6 @@ class UtilsUnitTest
   void _testStackTrace();
   void _testSetClassConfig();
   void _testFloatingException();
-  void _testEvents();
   void _testConvertFromString();
   void _testConvertFromStringToInt32Array();
   void _testCommandLine();
@@ -164,7 +154,6 @@ executeTest()
   _testUserData();
   _testHPReal();
   _testTruncateReal();
-  _testEvents();
 
   info() << Trace::Color::darkRed()     << "[TEST_COLOR] DARK_RED";
   info() << Trace::Color::darkGreen()   << "[TEST_COLOR] DARK_GREEN";
@@ -755,16 +744,6 @@ _printMemoryInfos()
     MemoryInfoPrinter mip(traceMng());
     mem_info->visitAllocatedBlocks(&mip);
   }
-}
-
-/*---------------------------------------------------------------------------*/
-/*---------------------------------------------------------------------------*/
-
-void UtilsUnitTest::
-_testEvents()
-{
-  info() << "INTERNAL TEST EVENT";
-  _internalTestEvent();
 }
 
 /*---------------------------------------------------------------------------*/
